+++ apps/desktop/src/content/core/index.ts
@@ -1,2 +1,4 @@
 export * from "./coreVocabularyEntries";
+export * from "./coreVocabularyManifest";
 export * from "./entries";
+export * from "./pilotCoreVocabularyEntries";